Type cleanups
added new make targets: pydocstyle_report diff_pydocstyle_report

+++ src/toil/serviceManager.py
@@ -17,20 +17,20 @@ import logging
 import time
 from queue import Empty, Queue
 from threading import Event, Thread
-from typing import Dict, Iterable, Set, Optional
+from typing import Iterable, Set, Optional
 
 from toil.job import ServiceJobDescription
 from toil.lib.throttle import LocalThrottle, throttle
 from toil.jobStores.abstractJobStore import AbstractJobStore
 from toil.toilState import ToilState
 
-logger = logging.getLogger( __name__ )
+logger = logging.getLogger(__name__)
+
 
 class ServiceManager:
-    """
-    Manages the scheduling of services.
-    """
-    def __init__(self, job_store: AbstractJobStore, toil_state: ToilState):
+    """Manages the scheduling of services."""
+
+    def __init__(self, job_store: AbstractJobStore, toil_state: ToilState) -> None:
         logger.debug("Initializing service manager")
         self.__job_store = job_store
 
@@ -60,7 +60,8 @@ class ServiceManager:
         # This is the queue of services for the batch system to start
         self.__services_out: Queue[str] = Queue()
 
-        self.__service_manager_jobs = 0 # The number of jobs the service manager is scheduling
+        # The number of jobs the service manager is scheduling
+        self.__service_manager_jobs = 0
 
         # Set up the service-managing thread.
         self.__service_starter = Thread(target=self.__start_services, daemon=True)
@@ -75,24 +76,21 @@ class ServiceManager:
         """
         Get the total number of jobs we are working on (services and their parent non-service jobs).
         """
-
         return self.__service_manager_jobs
 
     def start(self) -> None:
-        """
-        Start the service scheduling thread.
-        """
+        """Start the service scheduling thread."""
         self.__service_starter.start()
 
     def put_client(self, client_id: str) -> None:
         """
         Schedule the services of a job asynchronously.
+
         When the job's services are running the ID for the job will
         be returned by toil.leader.ServiceManager.get_ready_client.
 
         :param client_id: ID of job with services to schedule.
         """
-
         # Go get the client's description, which includes the services it needs.
         client = self.__toil_state.get_job(client_id)
 
@@ -102,7 +100,9 @@ class ServiceManager:
         self.__waiting_clients.add(client_id)
 
         # Add number of jobs managed by ServiceManager
-        self.__service_manager_jobs += len(client.services) + 1 # The plus one accounts for the root job
+        self.__service_manager_jobs += (
+            len(client.services) + 1
+        )  # The plus one accounts for the root job
 
         # Asynchronously schedule the services
         self.__clients_in.put(client_id)
@@ -139,7 +139,7 @@ class ServiceManager:
 
     def get_startable_service(self, maxWait: float) -> Optional[str]:
         """
-        :param float maxWait: Time in seconds to wait to get a job before returning.
+        :param maxWait: Time in seconds to wait to get a job before returning.
         :return: the ID of a service job that the leader can start, or None if no such job exists.
         """
         try:
@@ -150,6 +150,14 @@ class ServiceManager:
         except Empty:
             return None
 
+    def _get_service_job(self, service_id: str) -> ServiceJobDescription:
+        service = self.__toil_state.get_job(service_id)
+        if not isinstance(service, ServiceJobDescription):
+            raise Exception(
+                f"Expected ServiceJobDescription, got '{type(service)}': {service}."
+            )
+        return service
+
     def kill_services(self, service_ids: Iterable[str], error: bool = False) -> None:
         """
         Stop all the given service jobs.
@@ -159,35 +167,41 @@ class ServiceManager:
         """
         for service_id in service_ids:
             # Get the job description, which knows about the flag files.
-            service = self.__toil_state.get_job(service_id)
-            assert isinstance(service, ServiceJobDescription)
+            service = self._get_service_job(service_id)
+            if service.errorJobStoreID is None or service.terminateJobStoreID is None:
+                raise Exception("Must be a registered ServiceJobDescription")
             if error:
                 self.__job_store.delete_file(service.errorJobStoreID)
             self.__job_store.delete_file(service.terminateJobStoreID)
 
     def is_active(self, service_id: str) -> bool:
         """
-        Returns true if the service job has not been told to terminate.
+        Return true if the service job has not been told to terminate.
 
         :param service_id: Service to check on
-        :rtype: boolean
         """
-        service = self.__toil_state.get_job(service_id)
+        service = self._get_service_job(service_id)
+        if service.terminateJobStoreID is None:
+            raise Exception("Must be a registered ServiceJobDescription")
         return self.__job_store.file_exists(service.terminateJobStoreID)
 
     def is_running(self, service_id: str) -> bool:
         """
-        Returns true if the service job has started and is active
+        Return true if the service job has started and is active.
 
         :param service: Service to check on
-        :rtype: boolean
         """
-        service = self.__toil_state.get_job(service_id)
-        return (not self.__job_store.file_exists(service.startJobStoreID)) and self.is_active(service_id)
+        service = self._get_service_job(service_id)
+        if service.startJobStoreID is None:
+            raise Exception("Must be a registered ServiceJobDescription")
+        return (
+            not self.__job_store.file_exists(service.startJobStoreID)
+        ) and self.is_active(service_id)
 
     def check(self) -> None:
         """
         Check on the service manager thread.
+
         :raise RuntimeError: If the underlying thread has quit.
         """
         if not self.__service_starter.is_alive():
@@ -195,8 +209,9 @@ class ServiceManager:
 
     def shutdown(self) -> None:
         """
-        Cleanly terminate worker threads starting and killing services. Will block
-        until all services are started and blocked.
+        Terminate worker threads cleanly; starting and killing all service threads.
+
+        Will block until all services are started and blocked.
         """
         logger.debug('Waiting for service manager thread to finish ...')
         start_time = time.time()
@@ -205,13 +220,13 @@ class ServiceManager:
         # Kill any services still running to avoid deadlock
         for services in list(self.__toil_state.servicesIssued.values()):
             self.kill_services(services, error=True)
-        logger.debug('... finished shutting down the service manager. Took %s seconds', time.time() - start_time)
+        logger.debug(
+            "... finished shutting down the service manager. Took %s seconds",
+            time.time() - start_time,
+        )
 
     def __start_services(self) -> None:
-        """
-        Thread used to schedule services.
-        """
-
+        """Thread used to schedule services."""
         # Keep the user informed, but not too informed, as services start up
         log_limiter = LocalThrottle(60)
 
@@ -241,14 +256,18 @@ class ServiceManager:
                         remaining_services_by_client[client_id] = len(batch)
                         for service_id in batch:
                             # Load up the service object.
-                            service_job_desc = self.__toil_state.get_job(service_id)
+                            service_job_desc = self._get_service_job(service_id)
                             # Remember the parent job
                             service_to_client[service_id] = client_id
                             # We should now start to monitor this service to see if
                             # it has started yet.
                             starting_services.add(service_id)
                             # Send the service JobDescription off to be started
-                            logger.debug('Service manager is starting service job: %s, start ID: %s', service_job_desc, service_job_desc.startJobStoreID)
+                            logger.debug(
+                                "Service manager is starting service job: %s, start ID: %s",
+                                service_job_desc,
+                                service_job_desc.startJobStoreID,
+                            )
                             self.__services_out.put(service_id)
                 except Empty:
                     # No new jobs that need services scheduled.
@@ -259,18 +278,31 @@ class ServiceManager:
                     logger.debug('%d services are starting...', pending_service_count)
 
                 for service_id in list(starting_services):
-                    service_job_desc = self.__toil_state.get_job(service_id)
+                    service_job_desc = self._get_service_job(service_id)
+                    if (
+                        service_job_desc.startJobStoreID is None
+                        or service_job_desc.errorJobStoreID is None
+                    ):
+                        raise Exception("Must be a registered ServiceJobDescription")
                     if not self.__job_store.file_exists(service_job_desc.startJobStoreID):
                         # Service has started (or failed)
-                        logger.debug('Service %s has removed %s and is therefore started', service_job_desc, service_job_desc.startJobStoreID)
+                        logger.debug(
+                            "Service %s has removed %s and is therefore started",
+                            service_job_desc,
+                            service_job_desc.startJobStoreID,
+                        )
                         starting_services.remove(service_id)
                         client_id = service_to_client[service_id]
                         remaining_services_by_client[client_id] -= 1
                         assert remaining_services_by_client[client_id] >= 0
                         del service_to_client[service_id]
                         if not self.__job_store.file_exists(service_job_desc.errorJobStoreID):
-                            logger.error('Service %s has immediately failed before it could be used', service_job_desc)
-                            # It probably hasn't fileld in the promise that the job that uses the service needs.
+                            logger.error(
+                                "Service %s has immediately failed before it could be used",
+                                service_job_desc,
+                            )
+                            # It probably hasn't fileld in the promise that the
+                            # job that uses the service needs.
                             clients_with_failed_services.add(client_id)
 
                 # Find if any clients have had *all* their services started.
@@ -292,7 +324,6 @@ class ServiceManager:
         Wait until all the services for the given job are started, starting
         them in batches that are all issued together.
         """
-
         # Keep the user informed, but not too informed, as services start up
         log_limiter = LocalThrottle(60)
 
@@ -303,10 +334,23 @@ class ServiceManager:
             wait_on = []
             for service_id in service_job_list:
                 # Find the service object.
-                service_job_desc = self.__toil_state.get_job(service_id)
-                logger.debug("Service manager is starting service job: %s, start ID: %s", service_job_desc, service_job_desc.startJobStoreID)
-                assert self.__job_store.file_exists(service_job_desc.startJobStoreID), f"Service manager attempted to start service {service_job_desc} that has already started"
-                assert self.__toil_state.job_exists(service_job_desc.jobStoreID), f"Service manager attempted to start service {service_job_desc} that is not in the job store"
+                service_job_desc = self._get_service_job(service_id)
+                if (
+                    service_job_desc.startJobStoreID is None
+                    or service_job_desc.jobStoreID is None
+                ):
+                    raise Exception("Must be a registered ServiceJobDescription")
+                logger.debug(
+                    "Service manager is starting service job: %s, start ID: %s",
+                    service_job_desc,
+                    service_job_desc.startJobStoreID,
+                )
+                assert self.__job_store.file_exists(
+                    service_job_desc.startJobStoreID
+                ), f"Service manager attempted to start service {service_job_desc} that has already started"
+                assert self.__toil_state.job_exists(
+                    str(service_job_desc.jobStoreID)
+                ), f"Service manager attempted to start service {service_job_desc} that is not in the job store"
                 # At this point the terminateJobStoreID and errorJobStoreID
                 # could have been deleted, since the service can be killed at
                 # any time! So we can't assert their presence here.
@@ -317,7 +361,9 @@ class ServiceManager:
             # Wait until all the services of the batch are running
             for service_id in service_job_list:
                 # Find the service object.
-                service_job_desc = self.__toil_state.get_job(service_id)
+                service_job_desc = self._get_service_job(service_id)
+                if service_job_desc.startJobStoreID is None:
+                    raise Exception("Must be a registered ServiceJobDescription")
                 while self.__job_store.file_exists(service_job_desc.startJobStoreID):
                     # Sleep to avoid thrashing
                     time.sleep(1.0)
@@ -329,8 +375,11 @@ class ServiceManager:
                     if self.__terminate.is_set():
                         return
 
-                    if (not self.__toil_state.job_exists(service_job_desc.jobStoreID)
-                        and self.__job_store.file_exists(service_job_desc.startJobStoreID)):
+                    if not self.__toil_state.job_exists(
+                        str(service_job_desc.jobStoreID)
+                    ) and self.__job_store.file_exists(
+                        service_job_desc.startJobStoreID
+                    ):
                         # The service job has gone away but the service never flipped its start flag.
                         logger.error('Service %s has completed and been removed without ever starting', service_job_desc)
                         # Stop waiting on the service because we know it won't
